+++ CHANGELOG.md
@@ -5,5 +5,5 @@
 ### Added
 
 - `Ease` utilities for interpolation
-- Tween API using `EntityManager`
+- Tween API using `EntityManager`, `EntityCommandBuffer` and `ParallelWriter`
 - `Translation` tween support